Fix sass deprecation warnings
Actions taken:

```
git@github.com:vhx/foundation-sites.git
cd foundation-sites
git checkout v5.5.2
git checkout -b v5.5.2-with-sass-deprecation-warnings-fixed
npm install -g sass-migrator
sass-migrator division **/*.scss
```

as recommended here:
https://sass-lang.com/documentation/breaking-changes/slash-div

Context:

we're not likely to invest in upgrading foundation-sites because we are
migrating off of it. But in the meantime, it would be nice to address
the warnings we're seeing with our older version of it.